Move more code.
