+++ crates/core/src/lib.rs
@@ -5,7 +5,6 @@ mod host_funcs;
 mod proto;
 mod proto_config;
 mod shim_registry;
-mod shimmer;
 mod tool;
 mod tool_loader;
 mod tool_manifest;
@@ -20,7 +19,6 @@ pub use helpers::*;
 pub use proto::*;
 pub use proto_config::*;
 pub use semver::{Version, VersionReq};
-pub use shimmer::*;
 pub use tool::*;
 pub use tool_loader::*;
 pub use tool_manifest::*;
